==> code/1_taxonomic_assignment_blastn.Rmd <==
---
title: "Taxonomic Assignment"
author: "Kimberly Ledger"
date: "2025-03-26"
output: html_document
---

taxonomic assignment of preprocessed ASVs (dadasnake/DBO_mifish_20250325/post/filtered.seqs.fasta)

blastn -db ${DB}/nt -query ${QUERY} -perc_identity 96 -qcov_hsp_perc 98 -num_threads 10 -out blastn_out -outfmt '6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore sscinames staxids'

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load-libraries}
library(tidyverse)
library(tidyr)
library(stringr)
rename <- dplyr::rename
```

## load output from taxonkit 
```{r load-output-from-taxonkit}
taxonomy <-read.delim("/home/kimberly.ledger/dbo-mifish-mb/data/blastn_taxlineage.txt", header = FALSE, na.strings=c(""," ","NA"))
head(taxonomy)
```


## clean up the header a bit
```{r}
# use the full taxonomy rather than the seq id to collapse identical entries
tax_df <- taxonomy %>%
  filter(V4 > 100) %>% # make sure all retained matches are >100 bp
  select(-V2, -V5, -V6, -V7, -V14, -V8, -V9, -V10, -V11, -V12, -V13, -V15, -V16) %>%  #remove unnecessary columns
  group_by(V1, V17) %>% # group by the sequence key and the full taxonomy to reduce duplicate entries
  unique() %>% # doing that reduced the number of entries
  rename(qseqid=V1, perc_id=V3, length=V4, taxonomy=V17) %>% #rename headers
  filter(!is.na(taxonomy))
```

## calcualate the number of ASV's with taxonomic assignment
```{r how-many-total-ASVs-w-taxonomy}
tax_df %>% 
  ungroup() %>%
  select(qseqid) %>%
  unique() %>%
  nrow
```

## reformat table 
```{r}
# formatting the taxonomy variables
taxon_df <- tax_df %>%
  filter(str_detect(taxonomy, ";")) %>%
  separate(taxonomy, into=c("kingdom", "phylum", "class", "order", "family", "genus", "species"), sep = ";") %>% 
  filter(!grepl("sp\\.", species, ignore.case = TRUE)) %>%
  filter(!grepl("cf\\.", species, ignore.case = TRUE)) %>%
  filter(!grepl("aff\\.", species, ignore.case = TRUE)) %>%
  filter(!grepl(" x ", species, ignore.case = TRUE)) %>%         ### remove hybrids and unknown species assignments
  filter(!grepl("\\d", species)) %>%
  filter(!grepl("\\uncultured", species)) %>%
  filter(!grepl("\\unidentified", species)) %>%
  filter(!grepl("bacterium", species)) %>%
  filter(!grepl("var\\.", species, ignore.case = TRUE)) %>% 
  filter(!grepl("synthetic", species)) 
```

## what non-fish are here? 
```{r}
not_Actinopteri <- taxon_df %>%
  filter(class != "Actinopteri") 

unique(not_Actinopteri$species)
```


## manually clean up the non-fish

### figure out what ASVs are unambiguously a single species
```{r}
nf_single_spp_seqs <- not_Actinopteri %>% 
  group_by(qseqid) %>%
  add_tally(name = "n_taxa") %>%
  filter(n_taxa == 1) %>%
  filter(perc_id > 97)
```  
  
## remove the single-species seqs from the dataframe and then rank the hits by % identity for the remaining seqs
```{r}
nf_seq_id_diff <- not_Actinopteri %>%
  filter(perc_id > 97) %>%
  anti_join(., nf_single_spp_seqs) %>%  ## comment out this line if there are no single species ASVs
  select(-length) %>%
  group_by(qseqid, species, genus, family, order, class, phylum, kingdom) %>%
    mutate(seq_percID = max(perc_id)) %>%
    group_by(qseqid, species, genus, family, order, class, phylum, kingdom, seq_percID) %>%
  summarise(max(seq_percID)) %>% # take just the top hit for each taxon (for each sequence)
  select(-`max(seq_percID)`) %>%
  ungroup() %>%
  group_by(qseqid) %>%
      #mutate(id_rank = rank(desc(seq_percID), ties.method = "min")) %>% # rank the taxonomic hits per sequence by % id
       mutate(top_perc = max(seq_percID)) %>% # designate the highest % id for the best taxonomic hit in each sequence (in some, but not all cases, this is 100%)   %>%
      mutate(diff = top_perc - seq_percID) %>% # calculate the difference between the % identity of the top hit and each subsequent taxonomic hit
        arrange(diff)
```

Now you have the single best entry for each species for each sequence ranked and with the difference between the first and second ranked entries calculated.
** if there is a sequence that is at least a 99% match, then keep just the sequences within 0.5%  
** if there is NOT at least a 99% match, keep the top 1% 

### ASVs with 99% match or not 
```{r}
nf_match_99 <- nf_seq_id_diff %>%
  filter(top_perc >= 99) %>%
  select(qseqid) %>%
  unique()

nf_match_lessthan99<- nf_seq_id_diff %>%
  filter(top_perc < 99) %>%
  select(qseqid) %>%
  unique()
```

### filter based on percent identity 
```{r}
nf_match_99_to_remove_low_perc_hits <- nf_seq_id_diff %>%
  ungroup() %>%
  filter(qseqid %in% nf_match_99$qseqid) %>%
  group_by(qseqid) %>%
  filter(diff > 0.5)

nf_match_lessthan99_to_remove_low_perc_hits <- nf_seq_id_diff %>%
  ungroup() %>%
  filter(qseqid %in% nf_match_lessthan99$qseqid) %>%
  group_by(qseqid) %>%
  filter(diff > 1)

nf_to_remove_low_perc_hits <- nf_match_99_to_remove_low_perc_hits %>%
  bind_rows(nf_match_lessthan99_to_remove_low_perc_hits)

nf_keepers <- nf_seq_id_diff %>%
  anti_join(nf_to_remove_low_perc_hits)
```

after manually looking through the assignments, removing Pusa sibirica 
```{r}
nf_keepers <- nf_keepers %>%
  filter(!species == "Pusa sibirica")

nf_singletons <- nf_keepers %>%
  select(qseqid) %>%
  tally() %>%
  filter(n == 1)

# these are the seqs that now have only a single match
nf_singleton_df <- nf_singletons %>%
  left_join(nf_keepers) %>%
  select(-n) %>%
  bind_rows(nf_single_spp_seqs) %>% # combine the single spp data
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species)

## Genus-level matches
# remove the singletons from the bigger df 
nf_single_genus <- nf_keepers %>%
  anti_join(nf_singleton_df)%>% # at best, these should be genus-level matches
  group_by(qseqid, genus) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single genus
  select(-n) %>%
  left_join(., nf_keepers) %>%
  mutate(taxonomic_level = "genus") %>%
  mutate(taxon = genus)

## Family-level matches
nf_single_family <- nf_keepers %>%
  anti_join(nf_singleton_df)%>%
  anti_join(nf_single_genus) %>%
  group_by(qseqid, family) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single family
  select(-n) %>%
  left_join(., nf_keepers) %>%
  mutate(taxonomic_level = "family") %>%
  mutate(taxon = family)
```

## combine the non-fish assignments 
```{r}
nf_single_spp <- nf_singleton_df %>%
  select(-perc_id, -length, -n_taxa) %>%
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species)

nf_taxon_df_filter <- bind_rows(nf_single_genus, nf_single_spp, nf_single_family)
```


# check the fishes

## now look at the fish and figure out what fish taxa are in/out of our range
```{r}
to_check_range <- taxon_df %>%
  filter(class == "Actinopteri") %>%
  ungroup() %>%
  select(species, genus, family, order, class, phylum) %>%
  unique()
```

## check ranges for species using rfishbase 
```{r}
#library(remotes)
#remotes::install_github("ropensci/rfishbase")
library(rfishbase)

#one-time download of all fishbase tables... this takes a bit 
#fb_import()

#first, validate species names using rfishbase synonyms
spp_df <- synonyms(to_check_range$species)

syn <- spp_df %>% 
  filter(Status == "synonym")

#rename any synonyms with the validated name  
to_check_range_validated <- to_check_range %>%
  rename(synonym = species) %>%
  left_join(syn, by = "synonym") %>%
  mutate(synonym = ifelse(is.na(Species), synonym, Species)) %>%
  select(synonym:phylum) %>%
  rename(species = synonym)

#get distribution info 
spp_distribution <- distribution(to_check_range_validated$species) %>%
  select(Species, FAO) %>%
  unique()

#add column to designate if we will consider a species as "in range"- for this study, this will be NE Pacific 
spp_distribution <- spp_distribution %>%
  mutate(in_range = ifelse(is.na(FAO), NA, "no"),
         in_range = ifelse(FAO == "Pacific, Northeast", "yes", in_range),
         in_range = ifelse(FAO == "Arctic Ocean", "yes", in_range))

#keep just a list of spp names and yes/no/NA for "in range"  - this way we can keep track of what spp didn't have any reference information in fishbase to determine range 
spp_range <- spp_distribution %>%
  select(Species, in_range) %>%
  unique()

#how many entries do not have range info? 
range_na <- spp_range %>%
  filter(is.na(in_range))
```

When a valid name was not found, the presence of a species in the study area was checked using the GBIF database (https://www.gbif.org/).

## some species do not have range info - manually determine if these species should be considered in range
```{r}
spp_range <- spp_range %>%
  mutate(in_range = ifelse(Species == "Sebastes cheni", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Ammodytes japonicus", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Bothrocara zestum", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Icelus mororanis", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Sebastiscus vibrantus", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Phoxinus fayollarum", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Phoxinus dragarum", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Hadropogonichthys leptopus", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Platycaranx chrysophrys", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Barbatula hispanica", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Barbatula leoparda", "no", in_range))

spp_range %>%
  filter(is.na(in_range))
```


- in addition to fishbase distributions, these citations may help: https://www.marinespecies.org/arms/aphia.php?p=taxdetails%26id=10194
and exclude Atlantic Arctic species: Mecklenburg, C. W., Lynghammar, A., Johannesen, E., Byrkjedal, I., Christiansen, J. S., Dolgov, A. V., Karamushko, O. V., Mecklenburg, T. A., Møller, P. R., Steinke, D., & Wienerroither, R. M. (2018). Marine fishes of the Arctic region. Conservation of Arctic Flora and Fauna.

```{r}
atlantic_arctic_spp <- read.csv("/home/kimberly.ledger/BeringSea_Arctic_eDNA/data/AtlanticArctic_SpeciesList.csv")

in_range <- spp_range %>%
  filter(in_range == "yes") %>%
  filter(!Species %in% atlantic_arctic_spp$Name) %>%
  mutate(in_range = ifelse(Species == "Seriola lalandi", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Sardinops melanostictus", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Alosa sapidissima", "no", in_range)) 
```


```{r}
## what species are in range but used a synonym? 
syn_used <- in_range %>%
  filter(in_range == "yes") %>%
  filter(Species %in% syn$Species)
syn_used 

#what names were used in the tax_df? 
syn %>%
  filter(Species %in% syn_used$Species)

#rename these in the sorted_tax_df
taxon_df_renamed <- taxon_df %>%
  mutate(species = ifelse(species == "Gadus ogac", "Gadus macrocephalus", species), #
         #taxon = ifelse(taxon == "Gadus ogac", "Gadus macrocephalus", taxon), #
         species =  ifelse(species == "Embassichthys bathybius", "Microstomus bathybius", species), #
         #taxon = ifelse(taxon == "Embassichthys bathybius", "Microstomus bathybius", taxon), #
         genus = ifelse(species == "Microstomus bathybius", "Microstomus", genus), #
         species =  ifelse(species == "Ulcina olrikii", "Aspidophoroides olrikii", species), #
         #taxon = ifelse(taxon == "Ulcina olrikii", "Aspidophoroides olrikii", taxon), #
         genus = ifelse(species == "Aspidophoroides olrikii", "Aspidophoroides", genus), #
         species = ifelse(species == "Polypera greeni", "Liparis greeni", species), #
         #taxon = ifelse(taxon == "Polypera greeni", "Liparis greeni", taxon), #
         genus = ifelse(species == "Liparis greeni", "Liparis", genus), #
         species = ifelse(species == "Lycodopsis pacificus", "Lycodes pacificus", species), #
         #taxon = ifelse(taxon == "Lycodopsis pacificus", "Lycodes pacificus", taxon), #
         genus = ifelse(species == "Lycodes pacificus", "Lycodes", genus)) #
```


## keep only the "in range" fish species in the taxon dataframe - and all non-fish for now
```{r}
taxon_df_reduced <- taxon_df_renamed %>%
  filter(species %in% in_range$Species)
```


## figure out what ASVs are unambiguously a single species
```{r}
single_spp_seqs <- taxon_df_reduced %>% 
  group_by(qseqid) %>%
  add_tally(name = "n_taxa") %>%
  filter(n_taxa == 1) %>%
  filter(perc_id > 97)
```

## remove the single-species seqs from the dataframe and then rank the hits by % identity for the remaining seqs
```{r}
seq_id_diff <- taxon_df_reduced %>%
  anti_join(., single_spp_seqs) %>%  ## comment out this line if there are no single species ASVs
  select(-length) %>%
  group_by(qseqid, species, genus, family, order, class, phylum, kingdom) %>%
    mutate(seq_percID = max(perc_id)) %>%
    group_by(qseqid, species, genus, family, order, class, phylum, kingdom, seq_percID) %>%
  summarise(max(seq_percID)) %>% # take just the top hit for each taxon (for each sequence)
  select(-`max(seq_percID)`) %>%
  ungroup() %>%
  group_by(qseqid) %>%
      mutate(id_rank = rank(desc(seq_percID), ties.method = "min")) %>% # rank the taxonomic hits per sequence by % id
      mutate(top_perc = max(seq_percID)) %>% # designate the highest % id for the best taxonomic hit in each sequence (in some, but not all cases, this is 100%)   %>%
      mutate(diff = top_perc - seq_percID) %>% # calculate the difference between the % identity of the top hit and each subsequent taxonomic hit
        arrange(diff)
```

Now you have the single best entry for each species for each sequence ranked and with the difference between the first and second ranked entries calculated.

Now you have the single best entry for each species for each sequence ranked and with the difference between the first and second ranked entries calculated.
** if there is a sequence that is >99 % match, then keep just the sequences within 0.5%  
** if there is NOT >99% match, keep the top 1% 

## ASVs with 100% match 
```{r}
match_99 <- seq_id_diff %>%
  filter(top_perc >= 99) %>%
  select(qseqid) %>%
  unique()

match_lessthan99 <- seq_id_diff %>%
  filter(top_perc < 99) %>%
  select(qseqid) %>%
  unique()
```


## filter based on percent identity 
```{r}
match_99_to_remove_low_perc_hits <- seq_id_diff %>%
  ungroup() %>%
  filter(qseqid %in% match_99$qseqid) %>%
  group_by(qseqid) %>%
  filter(diff > 0.5)

match_lessthan99_to_remove_low_perc_hits <- seq_id_diff %>%
  ungroup() %>%
  filter(qseqid %in% match_lessthan99$qseqid) %>%
  group_by(qseqid) %>%
  filter(diff > 1)

to_remove_low_perc_hits <- match_99_to_remove_low_perc_hits %>%
  bind_rows(match_lessthan99_to_remove_low_perc_hits)

keepers <- seq_id_diff %>%
  anti_join(to_remove_low_perc_hits)

unique(keepers$species)
```

## so now determine whether ASVs should be assigned to genus, family, order, etc. 
```{r}
singletons <- keepers %>%
  select(qseqid) %>%
  tally() %>%
  filter(n == 1)

# these are the seqs that now have only a single match
singleton_df <- singletons %>%
  left_join(keepers) %>%
  select(-n) %>%
  bind_rows(single_spp_seqs) %>% # combine the single spp data
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species)

## Genus-level matches
# remove the singletons from the bigger df 
single_genus <- keepers %>%
  anti_join(singleton_df)%>% # at best, these should be genus-level matches
  group_by(qseqid, genus) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single genus
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "genus") %>%
  mutate(taxon = genus)
  

## Family-level matches
single_family <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  group_by(qseqid, family) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single family
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "family") %>%
  mutate(taxon = family)


## Order-level matches
single_order <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  group_by(qseqid, order) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single order
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "order") %>%
  mutate(taxon = order)


## Class-level matches
single_class <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  anti_join(single_order) %>%
  group_by(qseqid, class) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>% 
  filter(n == 1) %>% # seqs that match a single class
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "class") %>%
  mutate(taxon = class)


## Phylum-level matches
single_phylum <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  anti_join(single_order) %>%
  anti_join(single_class) %>%
  group_by(qseqid, phylum) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>% 
  filter(n == 1) %>% # seqs that match a single phylum
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "phylum") %>%
  mutate(taxon = phylum)

# Kingdom level matches
single_kingdom <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  anti_join(single_order) %>%
  anti_join(single_class) %>%
  group_by(qseqid, kingdom) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>% 
  filter(n == 1) %>% # seqs that match a single kingdom
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "kingdom") %>%
  mutate(taxon = kingdom)
```

## modify the singleton_df to include the right variable headers
```{r}
single_spp <- singleton_df %>%
  select(-perc_id, -length, -n_taxa) %>%
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species)
```

## recombine the full data set now that the appropriate level of taxonomy has been determined
```{r}
taxon_df_filter <- bind_rows(single_kingdom, single_phylum, single_class, single_order, single_family, single_genus, single_spp)
```

## recombine with non-fish 
```{r}
taxon_df_all <- bind_rows(nf_taxon_df_filter, taxon_df_filter)
```

## double check the asvs that did not get a taxonomic assignment 
```{r}
removed <- taxon_df %>%
  filter(!qseqid %in% taxon_df_all$qseqid)
```


## Create output taxonomy data frames
```{r}
# and then collapse that down to just a single taxon per ASV
# can i collapse while keeping the taxonomy for the taxon....? 
taxon_df_collapsed <- taxon_df_all %>%
  mutate(species = ifelse(taxonomic_level == "family", NA, species),
         genus = ifelse(taxonomic_level == "family", NA, genus),
         species = ifelse(taxonomic_level == "genus", NA, species),
         family = ifelse(taxonomic_level == "order", NA, family),
         genus = ifelse(taxonomic_level == "order", NA, genus),
         species = ifelse(taxonomic_level == "order", NA, species)) %>%
  select(qseqid, taxon, taxonomic_level, species, genus, family, order, class, phylum, kingdom) %>%
  unique()
```

```{r}
write.csv(taxon_df_filter, "/home/kimberly.ledger/dbo-mifish-mb/outputs/taxonomy_20250610_uncollapsed.csv")
write.csv(taxon_df_collapsed, "/home/kimberly.ledger/dbo-mifish-mb/outputs/taxonomy_20250610_collapsed.csv")
```
