==> code/3_taxon_summary.Rmd <==
---
title: "Additional processing and data summary of MiFish Amplicons"
author: "Kimberly Ledger"
output: html_document
date: "2025-06-10"
---

libraries
```{r}
library(tidyverse)
library(stringr)
library(reshape2)
rename <- dplyr::rename
```

load the lulu curated asv table 
```{r}
asv_table <- read.csv("/home/kimberly.ledger/dbo-mifish-mb/ref/asvtable_lulu98.csv") 

#transpose 
asv_table <- data.frame(t(asv_table))

#set column names to be ASV# 
colnames(asv_table) <- asv_table[1,]

#remove row that has ASV#
asv_table <- asv_table[-1, ]

#make sure reads are numbers
# Convert all character columns to numeric
for (col in names(asv_table)) {
  asv_table[[col]] <- as.numeric(asv_table[[col]])
}

asv_table$sample_ID <- rownames(asv_table)  #make make sample ID a column 

#reorder columns and change NTC names to remove extra "_"
asv_table <- asv_table %>%
  select(sample_ID, everything()) 

#lulu output changed "_" to "."
asv_table$sample_ID <- gsub("\\.", "_", asv_table$sample_ID) 
```

filter this table to only include the ABL DBO 2023 and 2024 libraries 
```{r}
abl_libraries <- read.csv("/home/kimberly.ledger/dbo-mifish-mb/data/DBO2020to2024_metadata.csv") %>%
  filter(lab == "ABL") %>% 
  mutate(sample_ID = gsub("-", "_", library))

asv_table_filter <- asv_table %>%
  filter(sample_ID %in% abl_libraries$sample_ID)
```


pivot table longer 
```{r}
# make a variable for the first and last ASV column in the table
asv_first <- which(colnames(asv_table_filter) == "ASV_0001")  #this is the first asv column in this dataframe
asv_last <- ncol(asv_table_filter)

asv_table_long <- asv_table_filter %>%
  pivot_longer(cols = c(asv_first:asv_last), names_to = "ASV", values_to = "reads") %>%
  mutate(reads = as.numeric(reads)) %>%
  mutate(reads = ifelse(is.na(reads), 0, reads))

length(unique(asv_table_long$ASV))
```

join the taxonomic assignments
```{r}
taxonomy <- read.csv("/home/kimberly.ledger/dbo-mifish-mb/outputs/taxonomy_20250610_collapsed.csv") %>%
  rename(ASV = qseqid)
```

join
```{r}
join_table <- asv_table_long %>%
  left_join(taxonomy, by = "ASV") %>%
  filter(!is.na(taxon)) %>%
  filter(reads > 0) 

unique(join_table$taxon)

join_table %>%
  filter(class != "Actinopteri") %>%
  ungroup() %>%
  select(class, family, taxon) %>%
  distinct(class, family, taxon)
```

remove the non-target species - non-marine mammals and birds
- note: Acipenser (sturgeon) was used as the positive control - keeping it in the df for now
```{r}
non_target <- c("Homo sapiens", "Canis lupus", "Bos", "Sus scrofa", "Hominidae", "Phasianidae", "Ovis")

join_table_filtered <- join_table %>%
  filter(!taxon %in% non_target)
```

calculate the output read count, number of asvs, and number of taxa assigned per sample replicate 
```{r}
output_summary <- join_table_filtered %>%
  group_by(sample_ID) %>%
  summarize(output_read_count = sum(reads),
            output_asv_count = n_distinct(ASV),
            output_num_tax_assigned = n_distinct(taxon))
```

join to full sampleID list to fill in the samples with no reads 
```{r}
output_summary_full <- abl_libraries %>%
  select(sample_ID) %>%
  left_join(output_summary) %>%
  mutate(across(everything(), ~replace_na(., 0)))
```

```{r}
write.csv(output_summary_full, "/home/kimberly.ledger/dbo-mifish-mb/outputs/output_read_summary.csv")
```

output a taxon table 
```{r}
taxon_table <- join_table_filtered %>%
  group_by(sample_ID, taxon, taxonomic_level, species, genus, family, order, class, kingdom) %>%
  summarize(reads = sum(reads))
```

quick summary
```{r}
summary_tb <- taxon_table %>%
  select(taxon:class, reads) %>% 
  group_by(taxon, taxonomic_level, species, genus, family, order, class) %>%
  summarize(total_reads = sum(reads)) %>% 
  arrange(desc(total_reads))
```


output taxon table 
```{r}
write.csv(taxon_table, "/home/kimberly.ledger/dbo-mifish-mb/outputs/taxon_table.csv")
```
